Now using strict rank and dtype promotions when testing.

+++ tests/test_while_loop.py
@@ -36,8 +36,8 @@ def _get_problem(key, *, num_steps: Optional[int]):
             step, val1, val2 = carry
             (theta,) = mlp(val1)
             real, imag = val1
-            z = real + imag * 1j
-            z = z * jnp.exp(1j * theta)
+            z = real.astype(jnp.complex64) + imag.astype(jnp.complex64) * 1j
+            z = z * jnp.exp(1j * theta.astype(jnp.complex64))
             real = jnp.real(z)
             imag = jnp.imag(z)
             jax.debug.print("{}", step)  # pyright: ignore
@@ -535,8 +535,8 @@ def test_speed_grad_checkpointed_while(getkey):
             step, val = carry
             (theta,) = mlp(val)
             real, imag = val
-            z = real + imag * 1j
-            z = z * jnp.exp(1j * theta)
+            z = real.astype(jnp.complex64) + imag.astype(jnp.complex64) * 1j
+            z = z * jnp.exp(1j * theta.astype(jnp.complex64))
             real = jnp.real(z)
             imag = jnp.imag(z)
             return step + 1, jnp.stack([real, imag])
